remove legacy markers
